7831 - Changed the moduel name and form name. Added an access handler for the product order multistep form. Added return type, check conditions, translatable string. Removed unnecessary function in js
