--- web/modules/custom/multistep_form/multistep_form.links.menu.yml
@@ -1,11 +1,0 @@
-multistep_form.multistep_form_settings:
-  title: Multi Step config
-  parent: system.admin_config_system
-  route_name: multistep_form.multistep_form_settings
-  weight: 10
-
-multistep_form.multistep_form_settings_js:
-  title: Multi Step config js
-  parent: system.admin_config_system
-  route_name: multistep_form.multistep_form_settings_js
-  weight: 10
