--- web/modules/custom/multistep_form/multistep_form.routing.yml
@@ -1,23 +1,0 @@
-multistep_form.multistep_form:
-  path: 'multistep/form'
-  defaults:
-    _form: '\Drupal\multistep_form\Form\MultistepForm'
-    _title: 'MultiStep Form'
-  requirements:
-    _permission: 'access content'
-
-multistep_form.multistep_form_settings:
-  path: '/admin/config/system/multistep'
-  defaults:
-    _form: '\Drupal\multistep_form\Form\MultistepFormSettings'
-    _title: 'MultiStep Form Settings'
-  requirements:
-    _permission: 'administer site configuration'
-
-multistep_form.multistep_form_settings_js:
-  path: '/admin/config/system/multistepjs'
-  defaults:
-    _form: '\Drupal\multistep_form\Form\MultistepFormSettingsJs'
-    _title: 'MultiStep Form Settings Js'
-  requirements:
-    _permission: 'administer site configuration'
