+++ web/modules/custom/product_order/product_order.links.menu.yml
@@ -1,0 +1,11 @@
+product_order.product_order_multistep_form_settings:
+  title: Product Order Multi Step config
+  parent: system.admin_config_system
+  route_name: product_order.product_order_multistep_form_settings
+  weight: 10
+
+product_order.product_order_multistep_form_settings_js:
+  title: Product Order Multi Step config js
+  parent: system.admin_config_system
+  route_name: product_order.product_order_multistep_form_settings_js
+  weight: 10
