+++ web/modules/custom/product_order/product_order.module
@@ -3,7 +3,7 @@
 /**
  * Implements hook_theme().
  */
-function multistep_form_theme($existing, $type, $theme, $path): array {
+function product_order_theme($existing, $type, $theme, $path): array {
 
   return [
     'multistep_ui_table' => [
