+++ web/modules/custom/product_order/src/Element/MultistepUiTable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\multistep_form\Element;
+namespace Drupal\product_order\Element;
 
 use Drupal\Component\Utility\Html;
 use Drupal\Core\Render\Element;
